[CORE] Move shuffle stream JNI bridge to JniCommon
